NyquistBase in lib-nyquist-effects
